backend/requirements.txt
fastapi
uvicorn[standard]
coqui-tts
